requirements.txt
aniso8601==7.0.0
asgiref==3.4.0
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Django==3.2.4
django-bootstrap==0.2.4
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-forms-bootstrap==3.1.0
django-phonenumber-field==5.2.0
django-preventconcurrentlogins==0.8.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
graphene==2.1.8
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
importlib-metadata==2.1.1
kiwisolver==1.3.1
matplotlib==3.4.2
mysqlclient==2.0.3
nose==1.3.7
numpy==1.21.1
pandas==1.3.1
phonenumbers==8.12.26
Pillow==8.3.0
promise==2.3
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-magic==0.4.24
python-slugify==5.0.2
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
Rx==1.6.1
selenium==3.141.0
singledispatch==3.6.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
Unidecode==1.2.0
urllib3==1.26.6
wfastcgi==3.0.0
xlrd==2.0.1
xmltodict==0.12.0
zipp==3.4.1
